make Fallible Custom Mutator the default and implement it's handle_err method by default
