+++ custom_mutators/rust/example/src/lib.rs
@@ -1,54 +1,14 @@
 #![allow(unused_variables)]
 
-use custom_mutator::{export_mutator, fallible::FallibleCustomMutator, CustomMutator};
+use custom_mutator::{export_mutator, CustomMutator};
 use std::os::raw::c_uint;
 
 struct ExampleMutator;
 
 impl CustomMutator for ExampleMutator {
-    fn init(seed: c_uint) -> Self {
-        Self
-    }
+    type Error = ();
 
-    fn fuzz<'b, 's: 'b>(
-        &'s mut self,
-        buffer: &'b mut [u8],
-        add_buff: Option<&[u8]>,
-        max_size: usize,
-    ) -> Option<&'b [u8]> {
-        buffer.reverse();
-        Some(buffer)
-    }
-}
-
-struct OwnBufferExampleMutator {
-    own_buffer: Vec<u8>,
-}
-
-impl CustomMutator for OwnBufferExampleMutator {
-    fn init(seed: c_uint) -> Self {
-        Self {
-            own_buffer: Vec::new(),
-        }
-    }
-
-    fn fuzz<'b, 's: 'b>(
-        &'s mut self,
-        buffer: &'b mut [u8],
-        add_buff: Option<&[u8]>,
-        max_size: usize,
-    ) -> Option<&'b [u8]> {
-        self.own_buffer.reverse();
-        Some(self.own_buffer.as_slice())
-    }
-}
-
-struct FallibleMutatorExample;
-
-impl FallibleCustomMutator for FallibleMutatorExample {
-    type Error = i32;
-
-    fn init(seed: c_uint) -> Result<Self, i32> {
+    fn init(seed: c_uint) -> Result<Self, ()> {
         Ok(Self)
     }
 
@@ -57,16 +17,33 @@ impl FallibleCustomMutator for FallibleMutatorExample {
         buffer: &'b mut [u8],
         add_buff: Option<&[u8]>,
         max_size: usize,
-    ) -> Result<Option<&'b [u8]>, i32> {
-        Err(1)
+    ) -> Result<Option<&'b [u8]>, ()> {
+        buffer.reverse();
+        Ok(Some(buffer))
+    }
+}
+
+struct OwnBufferExampleMutator {
+    own_buffer: Vec<u8>,
+}
+
+impl CustomMutator for OwnBufferExampleMutator {
+    type Error = ();
+
+    fn init(seed: c_uint) -> Result<Self, ()> {
+        Ok(Self {
+            own_buffer: Vec::new(),
+        })
     }
 
-    fn handle_err(err: i32) {
-        if std::env::var("DEBUG").is_ok() {
-            eprintln!("Error in custom mutator: {}", err);
-            // might panic in debug mode
-        }
-        // not panicing here -> continue execution
+    fn fuzz<'b, 's: 'b>(
+        &'s mut self,
+        buffer: &'b mut [u8],
+        add_buff: Option<&[u8]>,
+        max_size: usize,
+    ) -> Result<Option<&'b [u8]>, ()> {
+        self.own_buffer.reverse();
+        Ok(Some(self.own_buffer.as_slice()))
     }
 }
